TSV Export should now work. It uses now the correct property names and not the type names anymore. Furthermore the celcius conversion had to be done from an int kelvin, aka kelvin multiplied by 100.

+++ GatewayApp/Backend/Plantmonitor.Server/Features/ImageStitching/PhotoStitcher.cs
@@ -26,7 +26,7 @@ public class PhotoStitcher(ILogger<IPhotoStitcher> logger) : IPhotoStitcher
         public string Name { get; init; } = "";
         public DateTime IrImageTime { get; set; }
         public DateTime VisImageTime { get; set; }
-        public float IrTemperatureInK { get; set; }
+        public int IrTemperatureInK { get; set; }
         public string Comment { get; init; } = "";
 
         public void Dispose()
